From 6 to 9 spore types

+++ src/spore.rs
@@ -1,5 +1,3 @@
-use std::usize;
-
 use crate::{configuration::NUMBER_OF_CONFIGS, vector::Vector};
 
 type Buckets<T> = Vec<Vec<T>>;
@@ -13,7 +11,7 @@ pub struct SporesState {
 // Uses Arrays instead of Vectors. Arrays are probably loaded to the CPU cache, while Vector has one more level of redirection.
 #[derive(Debug)]
 pub struct SporeConfigs {
-    pub repulsion_dists: [f32; NUMBER_OF_CONFIGS as usize],
-    pub force_factors: [f32; NUMBER_OF_CONFIGS as usize],
-    pub force_reaches: [f32; NUMBER_OF_CONFIGS as usize],
+    pub repulsion_dists: [f32; NUMBER_OF_CONFIGS],
+    pub force_factors: [f32; NUMBER_OF_CONFIGS],
+    pub force_reaches: [f32; NUMBER_OF_CONFIGS],
 }
